fix some bugs in bitmap and string & add support in userprog

+++ src/include/lib/kernel/bitmap.h
@@ -11,7 +11,7 @@ struct bitmap {
 };
 
 void bitmap_init(struct bitmap* btmap);
-int bitmap_scan_test(struct bitmap* btmap, uint32_t bit_idx);
+uint8_t bitmap_scan_test(struct bitmap* btmap, uint32_t bit_idx);
 int bitmap_scan(struct bitmap* btmap, uint32_t cnt);
 void bitmap_set(struct bitmap* btmap, uint32_t index, int8_t value);
